♻️ adjust entrypoint for creating atoms
