+++ packages/atom.io/data/src/struct.ts
@@ -2,7 +2,7 @@ import type * as AtomIO from "atom.io"
 import type { Store } from "atom.io/internal"
 import { IMPLICIT } from "atom.io/internal"
 
-import { createAtom, createSelector } from "atom.io/internal"
+import { createRegularAtom, createSelector } from "atom.io/internal"
 
 const capitalize = (str: string) => str[0].toUpperCase() + str.slice(1)
 
@@ -29,7 +29,7 @@ export function struct<
 		>}State`]: AtomIO.AtomToken<Struct[K]>
 	} = Object.keys(options.default).reduce((acc, key) => {
 		const atomName = options.key + capitalize(key) + `State`
-		acc[atomName] = createAtom(
+		acc[atomName] = createRegularAtom(
 			{
 				key: `${options.key}.${key}`,
 				default: options.default[key],
